Setup unit tests
