+++ .vscode/tasks.json
@@ -1,0 +1,37 @@
+{
+    "version": "2.0.0",
+    "tasks": [
+        {
+            "label": "Build & Run GradeBook (Debug)",
+            "command": "cmake -B build . && cd build && make -j 8 && cd gradebook && ./gradebook",
+            "type": "shell",
+            "args": [],
+            "problemMatcher": [
+                "$tsc"
+            ],
+            "presentation": {
+                "reveal": "always"
+            },
+            "group": {
+                "kind": "build",
+                "isDefault": true
+            }
+        },
+        {
+            "label": "Run GradeBook Unit Tests",
+            "command": "cmake -DTESTING=ON -B build . && cd build && make -j 8 && cd tests && ./gradebook_test",
+            "type": "shell",
+            "args": [],
+            "problemMatcher": [
+                "$tsc"
+            ],
+            "presentation": {
+                "reveal": "always"
+            },
+            "group": {
+                "kind": "test",
+                "isDefault": true
+            }
+        }
+    ]
+}
