+++ tests/assets/assignmentdata.csv
@@ -1,0 +1,11 @@
+assignment_id,name,due_date,class_id
+1,Chemical Elements,2023-12-01,1
+2,Nuclear Fission,2023-11-15,1
+3,Matrices,2023-11-13,2
+4,Graphing,2023-11-30,2
+5,Gravity Studies,2023-12-03,3
+6,NASA Visit,2023-11-09,3
+7,Shakespeare Book Review,2023-12-04,4
+8,Old English,2023-11-17,4
+9,Vikings Raids,2023-11-28,5
+10,Abraham Lincoln,2023-11-08,5
