+++ tests/assets/classdata.csv
@@ -1,0 +1,6 @@
+class_id,name,grade_level,teacher_id
+1,Chemistry,12,1
+2,Mathematics,12,2
+3,Physics,12,3
+4,English,12,4
+5,History,12,5
