+++ tests/assets/grades.csv
@@ -1,0 +1,101 @@
+assignment_id,student_id,score
+1,1,72
+2,1,26
+3,1,51
+4,1,3
+5,1,20
+6,1,22
+7,1,69
+8,1,13
+9,1,56
+10,1,28
+1,2,20
+2,2,23
+3,2,67
+4,2,41
+5,2,65
+6,2,22
+7,2,82
+8,2,36
+9,2,26
+10,2,94
+1,3,7
+2,3,8
+3,3,13
+4,3,22
+5,3,13
+6,3,57
+7,3,34
+8,3,27
+9,3,39
+10,3,97
+1,4,50
+2,4,46
+3,4,42
+4,4,95
+5,4,26
+6,4,59
+7,4,33
+8,4,97
+9,4,87
+10,4,85
+1,5,25
+2,5,54
+3,5,52
+4,5,33
+5,5,25
+6,5,28
+7,5,59
+8,5,71
+9,5,8
+10,5,85
+1,6,11
+2,6,22
+3,6,36
+4,6,39
+5,6,43
+6,6,20
+7,6,32
+8,6,25
+9,6,15
+10,6,23
+1,7,78
+2,7,93
+3,7,7
+4,7,50
+5,7,75
+6,7,5
+7,7,14
+8,7,42
+9,7,55
+10,7,79
+1,8,75
+2,8,29
+3,8,95
+4,8,82
+5,8,22
+6,8,74
+7,8,23
+8,8,77
+9,8,48
+10,8,8
+1,9,98
+2,9,73
+3,9,52
+4,9,74
+5,9,54
+6,9,33
+7,9,35
+8,9,56
+9,9,65
+10,9,93
+1,10,14
+2,10,16
+3,10,55
+4,10,17
+5,10,51
+6,10,30
+7,10,77
+8,10,82
+9,10,99
+10,10,85
