+++ tests/assets/studentdata.csv
@@ -1,0 +1,10 @@
+student_id,name,grade_level,gpa
+1,John Doe,12,3.8
+2,Luke Sutton,12,4.7
+3,Jayden Gibson,12,4.6
+4,Leon Porter,12,3.3
+5,William Edwards,12,3.2
+6,Morgan Bennett,12,3.4
+7,Zander Bradford,12,3.3
+8,Franklin Osborn,12,3.3
+9,Mauricio Orr,12,4.3
