+++ tests/assets/teacherdata.csv
@@ -1,0 +1,6 @@
+teacher_id,name
+1,Adam Bell
+2,Tommy Barrett
+3,Ollie Perry
+4,Elliot Hudson
+5,Freddie Harris
